requirements.txt
pytest
parametrize_from_file
pytest-reportlog
pytest-html
pymongo
cython
mkdocs-material==9.6.1
mdx-include >=1.4.1,<2.0.0
mkdocs-github-admonitions-plugin
mkdocs-redirects>=1.2.1,<1.3.0
typer == 0.15.3
pyyaml >=5.3.1,<7.0.0
pillow==11.1.0
cairosvg==2.7.1
mkdocstrings[python]==0.26.1
griffe-typingdoc==0.2.7
mkdocs-macros-plugin==1.3.7
markdown-include-variants==0.0.4
pymdown-extensions
dacite
tqdm
debugpy
pyfaktory
lizard
mongomock
pre-commit